# Conflicts:
#	nts.uk/uk.at/at.file/nts.uk.file.at.infra/src/main/java/nts/uk/file/at/infra/schedule/monthly/AsposeMonthlyWorkScheduleGenerator.java

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.app/src/main/java/nts/uk/ctx/at/shared/app/command/yearholidaygrant/GrantConditionCommand.java
@@ -18,7 +18,7 @@ public class GrantConditionCommand {
 	private int conditionNo;
 	
 	/* 条件値 */
-	private Integer conditionValue;
+	private Double conditionValue;
 	
 	/* 条件利用区分 */
 	private int useConditionAtr;
